+++ CHANGES.md
@@ -6,6 +6,7 @@
 - CRT: Separate logging APIs from `libc.a` to `liblog.a` **(Disruptive)**
 - CRT: Added `vgs_memcpy`, `vgs_memset` and `vgs_strlen` _(Using DMA)_
 - CRT: Added `vgs_strcmp`, `vgs_strncmp`, `vgs_strstr`, `vgs_strchr` and `vgs_strrchr` _(Not DMA)_
+- CRT: Added `vgs_isdigit`, `vgs_isupper`, `vgs_islower`, `vgs_isalpha`, `vgs_isalnum`, `vgs_toupper`, `vgs_tolower`
 - Other: Preparing for Unit Testing in CI
 
 ## Version 0.2.0